requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-yasg==1.21.7
python-decouple==3.8
whitenoise==6.6.0
psycopg2-binary==2.9.9
Faker==20.1.0
django-environ==0.11.2
